# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API CRUD States Cities Documentation",
        "description": "An API with CRUD of States and Cities",
        "contact": {
            "email": "tarcisio.saraujo@gmail.com"
        },
        "license": {
            "name": "The MIT License (MIT)",
            "url": "https://mit-license.org/"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/cities": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "Get a list of cities",
                "description": "Returns a list of cities",
                "operationId": "App\\Http\\Controllers\\Api\\CitieController::index",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cities"
                ],
                "summary": "Store new citie",
                "description": "Store citie data",
                "operationId": "App\\Http\\Controllers\\Api\\CitieController::store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Citie name",
                        "required": true
                    },
                    {
                        "name": "state_id",
                        "in": "query",
                        "description": "State id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/cities/{id}": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "Get citie information",
                "description": "Returns citie data",
                "operationId": "App\\Http\\Controllers\\Api\\CitieController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Citie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cities"
                ],
                "summary": "Update existing citie",
                "description": "Update citie data",
                "operationId": "App\\Http\\Controllers\\Api\\CitieController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Citie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Citie name",
                        "required": true
                    },
                    {
                        "name": "state_id",
                        "in": "query",
                        "description": "State id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cities"
                ],
                "summary": "Delete existing citie",
                "description": "Delete a citie data",
                "operationId": "App\\Http\\Controllers\\Api\\CitieController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Citie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/states": {
            "get": {
                "tags": [
                    "States"
                ],
                "summary": "Get a list of states",
                "description": "Returns a list of states",
                "operationId": "App\\Http\\Controllers\\Api\\StateController::index",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "States"
                ],
                "summary": "Store new state",
                "description": "Store state data",
                "operationId": "App\\Http\\Controllers\\Api\\StateController::store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "State name",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/states/{id}": {
            "get": {
                "tags": [
                    "States"
                ],
                "summary": "Get state information",
                "description": "Returns state data",
                "operationId": "App\\Http\\Controllers\\Api\\StateController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "State id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "States"
                ],
                "summary": "Update existing state",
                "description": "Update state data",
                "operationId": "App\\Http\\Controllers\\Api\\StateController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "State id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "State name",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "delete": {
                "tags": [
                    "States"
                ],
                "summary": "Delete existing state",
                "description": "Delete a state data",
                "operationId": "App\\Http\\Controllers\\Api\\StateController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "State id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "API CRUD States Cities",
            "description": "API Endpoints of CRUD States Cities"
        }
    ],
    "security": [
        []
    ]
}